feat: Add impostors know each other setting
Introduces a new game configuration option that allows impostors to see the names of other impostors. This is achieved by adding a new property `impostorsKnowEachOther` to `GameConfig` and updating the `Player` interface to include `otherImpostors`.

The UI has been updated to include this setting in the setup screen and display the identified impostors on the role reveal screen when enabled. A new category "Costeños" has also been added.